+++ grafana-plugin/src/pages/settings/tabs/ChatOps/ChatOps.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { AppRootProps } from '@grafana/data';
-import { HorizontalGroup, Icon } from '@grafana/ui';
+import { Alert, HorizontalGroup, Icon } from '@grafana/ui';
 import cn from 'classnames/bind';
 import { observer } from 'mobx-react';
 
@@ -9,6 +9,7 @@ import VerticalTabsBar, { VerticalTab } from 'components/VerticalTabsBar/Vertica
 import SlackSettings from 'pages/settings/tabs/ChatOps/tabs/SlackSettings/SlackSettings';
 import TelegramSettings from 'pages/settings/tabs/ChatOps/tabs/TelegramSettings/TelegramSettings';
 import { AppFeature } from 'state/features';
+import { WithStoreProps } from 'state/types';
 import { useStore } from 'state/useStore';
 import { withMobXProviderContext } from 'state/withStore';
 import LocationHelper from 'utils/LocationHelper';
@@ -21,7 +22,7 @@ export enum ChatOpsTab {
   Slack = 'Slack',
   Telegram = 'Telegram',
 }
-interface ChatOpsProps extends AppRootProps {}
+interface ChatOpsProps extends AppRootProps, WithStoreProps {}
 interface ChatOpsState {
   activeTab: ChatOpsTab;
 }
@@ -44,6 +45,11 @@ class ChatOpsPage extends React.Component<ChatOpsProps, ChatOpsState> {
 
   render() {
     const { activeTab } = this.state;
+    const { store } = this.props;
+
+    if (!this.isChatOpsConfigured() && store.isOpenSource()) {
+      return this.renderNoChatOpsBannerInfo();
+    }
 
     return (
       <div className={cx('root')}>
@@ -55,6 +61,29 @@ class ChatOpsPage extends React.Component<ChatOpsProps, ChatOpsState> {
         </div>
       </div>
     );
+  }
+
+  renderNoChatOpsBannerInfo() {
+    return (
+      <div className={cx('root')} data-testid="chatops-banner">
+        <Alert severity="warning" title="No ChatOps found">
+          ChatOps is disabled because no chat integration is enabled. See{' '}
+          <a href="https://grafana.com/docs/oncall/latest/open-source/#telegram-setup" target="_blank" rel="noreferrer">
+            Telegram
+          </a>{' '}
+          and{' '}
+          <a href="https://grafana.com/docs/oncall/latest/open-source/#slack-setup" target="_blank" rel="noreferrer">
+            Slack
+          </a>{' '}
+          docs for more information.
+        </Alert>
+      </div>
+    );
+  }
+
+  isChatOpsConfigured(): boolean {
+    const { store } = this.props;
+    return store.hasFeature(AppFeature.Slack) || store.hasFeature(AppFeature.Telegram);
   }
 
   handleChatopsTabChange(tab: ChatOpsTab) {